Third commit

+++ resources/views/posts/show.blade.php
@@ -43,26 +43,56 @@
               </div>
             </div>
 
-            <div class="row">
-              <div class="col-1">
+              <div class="row">
+                @guest <!-- si es guest no muestra los botones de editar y eliminar -->
+                  <div class="col-1">
+                  </div>
+                  <div class="col-4">
+                  </div>
+                  <div class="col-2">
+                  </div>
+                  <div class="col-4">
+                  </div>
+                  <div class="col-1">
+                  </div>
+                @else <!-- el usuario logueado puede editar y eliminar sus propios posts -->
+                  @if($post->user_id == Auth::user()->id)
+
+                    <div class="col-4">
+                    {!! Html::linkRoute('posts.edit', 'Editar', array($post->id), array('class'=>'btn btn-primary btn-block')) !!}
+                    </div>
+                    <div class="col-4">
+                      {!! Form::open(['route' => ['posts.destroy', $post->id], 'method' => 'DELETE']) !!}
+                      {!! Form::submit('Eliminar', ['class' => 'btn btn-danger btn-block']) !!}
+                      {!! Form::close()!!}
+                    </div>
+                    <div class="col-4">
+                      {!! Html::linkRoute('posts.index', 'Ver todas mis publicaciones', array($post->id), array('class'=>'btn btn-primary btn-block')) !!}
+                    </div>
+
+                  @else <!-- el usuario logueado no puede editar ni eliminar posts de otros -->
+                    <div class="col-1">
+                    </div>
+                    <div class="col-4">
+                    </div>
+                    <div class="col-2">
+                    </div>
+                    <div class="col-4">
+                    </div>
+                    <div class="col-1">
+                    </div>
+                  @endif
               </div>
-              <div class="col-4">
-                {!! Html::linkRoute('posts.edit', 'Editar', array($post->id), array('class'=>'btn btn-primary btn-block')) !!}
-              </div>
-              <div class="col-2">
-              </div>
-              <div class="col-4">
-                {!! Form::open(['route' => ['posts.destroy', $post->id], 'method' => 'DELETE']) !!}
-                {!! Form::submit('Eliminar', ['class' => 'btn btn-danger btn-block']) !!}
-                {!! Form::close()!!}
-              </div>
-              <div class="col-1">
-              </div>
-            </div>
+            @endguest
 
           </div>
         </div>
-      </div>
+        <div class="row">
+          <div class="col-8">
+            <h4 class="subtitle_h3_nolines"><a href="/blog/{{$post->slug}}">{{ url('blog/'.$post->slug) }}</a></h4>
+          </div>
+        </div>
+      </div> <!-- end of card header -->
 
       <div class="card-body">
         <p class="card-text postbody" style="text-align: justify">{{$post->body}}</p>
